Accept float for seek position [Fixes #1439] (#1446)
The MediaPlayerEntity defines the seek position as a `float`, but the
mass entity is defining it as an `int`. The type hint is ignored and
the value is passed through as a float, which then fails validation on
the server-side at `PlayerQueuesController.seek()`.

For now the simplest solution is to truncate the extra precision and
pass it through as an int.